improve long support in logical functions in imagej.process.function.*
This used to be revision r1624.

+++ imagej/src/main/java/imagej/process/function/binary/AndBinaryFunction.java
@@ -4,6 +4,6 @@ public class AndBinaryFunction implements BinaryFunction
 {
 	public double compute(double input1, double input2)
 	{
-		return ((int)input1) & ((int)input2);
+		return ((long)input1) & ((long)input2);
 	}
 }
