+++ imagej/src/main/java/imagej/process/function/binary/XorBinaryFunction.java
@@ -4,7 +4,7 @@ public class XorBinaryFunction implements BinaryFunction {
 
 	public double compute(double input1, double input2)
 	{
-		return ( ((int)input1) ^ ((int)input2) );
+		return ( ((long)input1) ^ ((long)input2) );
 	}
 
 }
